add new custom method to execute a full proposal
refactor some utility functions moved to utils folder
