+++ src/execution/index.ts
@@ -2,3 +2,4 @@ import "./signing"
 import "./simple"
 import "./submitting"
 import "./proposing"
+import "./custom-execute"
