+++ src/execution/submitting.ts
@@ -1,79 +1,10 @@
-import { BigNumber, Contract, PopulatedTransaction, Signer, utils } from "ethers";
+import { BigNumber, PopulatedTransaction } from "ethers";
 import { task, types } from "hardhat/config";
 import { safeSingleton } from "../contracts";
-import { buildSafeTransaction, calculateSafeTransactionHash, populateExecuteTx, safeApproveHash, SafeSignature, SafeTransaction } from "@gnosis.pm/safe-contracts";
+import { buildSafeTransaction, populateExecuteTx } from "@gnosis.pm/safe-contracts";
 import { parseEther } from "@ethersproject/units";
-import { getAddress } from "@ethersproject/address";
 import { isHexString } from "ethers/lib/utils";
-import { SafeTxProposal } from "./proposing";
-import { loadSignatures, proposalFile, readFromCliCache } from "./utils";
-
-const parsePreApprovedConfirmation = (data: string): SafeSignature => {
-    const signer = getAddress("0x" + data.slice(26, 66))
-    return {
-        signer, data
-    }
-}
-
-const parseTypeDataConfirmation = (safeTxHash: string, data: string): SafeSignature => {
-    const signer = utils.recoverAddress(safeTxHash, data)
-    return {
-        signer, data
-    }
-}
-
-const parseEthSignConfirmation = (safeTxHash: string, data: string): SafeSignature => {
-    const signer = utils.recoverAddress(utils.hashMessage(utils.arrayify(safeTxHash)), data.replace(/1f$/, "1b").replace(/20$/, "1c"))
-    return {
-        signer, data
-    }
-}
-
-const parseSignature = (safeTxHash: string, signature: string): SafeSignature => {
-    if (!isHexString(signature, 65)) throw Error(`Unsupported signature: ${signature}`)
-    const type = parseInt(signature.slice(signature.length - 2), 16)
-    switch (type) {
-        case 1: return parsePreApprovedConfirmation(signature)
-        case 27:
-        case 28:
-            return parseTypeDataConfirmation(safeTxHash, signature)
-        case 31:
-        case 32:
-            return parseEthSignConfirmation(safeTxHash, signature)
-        case 0:
-        default:
-            throw Error(`Unsupported type ${type} in ${signature}`)
-    }
-}
-
-const isOwnerSignature = (owners: string[], signature: SafeSignature): SafeSignature => {
-    if (owners.indexOf(signature.signer) < 0) throw Error(`Signer ${signature.signer} not found in owners ${owners}`)
-    return signature
-}
-
-const prepareSignatures = async (safe: Contract, tx: SafeTransaction, signaturesCSV: string | undefined, submitter?: Signer, knownSafeTxHash?: string): Promise<SafeSignature[]> => {
-    const owners = await safe.getOwners()
-    const signatures = new Map<String, SafeSignature>()
-    const submitterAddress = submitter && await submitter.getAddress()
-    if (signaturesCSV) {
-        const chainId = (await safe.provider.getNetwork()).chainId
-        const safeTxHash = knownSafeTxHash ?? calculateSafeTransactionHash(safe, tx, chainId)
-        for (const signatureString of signaturesCSV.split(",")) {
-            const signature = isOwnerSignature(owners, parseSignature(safeTxHash, signatureString))
-            if (submitterAddress === signature.signer || signatures.has(signature.signer)) continue
-            signatures.set(signature.signer, signature)
-        }
-    }
-    const threshold = (await safe.getThreshold()).toNumber()
-    const submitterIsOwner = submitterAddress && owners.indexOf(submitterAddress) >= 0
-    const requiredSigntures = submitterIsOwner ? threshold - 1 : threshold
-    if (requiredSigntures > signatures.size) throw Error(`Not enough signatures (${signatures.size} of ${threshold})`)
-    const signatureArray = []
-    if (submitterIsOwner) {
-        signatureArray.push(await safeApproveHash(submitter!!, safe, tx, true))
-    }
-    return signatureArray.concat(Array.from(signatures.values()).slice(0, requiredSigntures))
-}
+import { loadSignatures, prepareSignatures, proposalFile, readFromCliCache, SafeTxProposal } from "./utils";
 
 task("submit-tx", "Executes a Safe transaction")
     .addPositionalParam("address", "Address or ENS name of the Safe to check", undefined, types.string)
